+++ fastmixture/svd.pyx
@@ -6,7 +6,7 @@ from libc.math cimport sqrt
 
 ##### Randomized SVD - PCAone method #####
 # Load centered chunk from PLINK file for SVD
-cpdef void plinkChunk(unsigned char[:,::1] G, float[:,::1] X, float[::1] f, \
+cpdef void plinkChunk(unsigned char[:,::1] G, double[:,::1] X, double[::1] f, \
 		int M_b, int t) nogil:
 	cdef:
 		int M = X.shape[0]
@@ -22,7 +22,7 @@ cpdef void plinkChunk(unsigned char[:,::1] G, float[:,::1] X, float[::1] f, \
 			byte = G[M_b+j,b]
 			for bytepart in range(4):
 				if recode[byte & mask] != 9:
-					X[j,i] = <float>recode[byte & mask] - 2.0*f[M_b+j]
+					X[j,i] = <double>recode[byte & mask] - 2.0*f[M_b+j]
 				else:
 					X[j,i] = 0.0
 				byte = byte >> 2
@@ -31,24 +31,24 @@ cpdef void plinkChunk(unsigned char[:,::1] G, float[:,::1] X, float[::1] f, \
 					break
 
 # Root-mean square error between two Q matrices
-cpdef float rmse(float[:,::1] A, float[:,::1] B) nogil:
+cpdef double rmse(double[:,::1] A, double[:,::1] B) nogil:
 	cdef:
 		int N = A.shape[0]
 		int K = A.shape[1]
 		int i, k
-		float res = 0.0
+		double res = 0.0
 	for i in range(N):
 		for k in range(K):
 			res += (A[i,k] - B[i,k])*(A[i,k] - B[i,k])
-	return sqrt(res/<float>(N*K))
+	return sqrt(res/<double>(N*K))
 
 # Map2domain
-cpdef void map2domain(float[:,::1] Q) nogil:
+cpdef void map2domain(double[:,::1] Q) nogil:
 	cdef:
 		int N = Q.shape[0]
 		int K = Q.shape[1]
 		int i, k
-		float sumQ
+		double sumQ
 	for i in range(N):
 		sumQ = 0.0
 		for k in range(K):
